+++ app.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+# Import routes
+import pyris.routes.llm
