+++ pyris/routes/llm.py
@@ -1,11 +1,8 @@
-from flask import Flask
-from models.dtos import SendMessageRequest, SendMessageResponse
+from app import app
+from pyris.models.dtos import SendMessageRequest, SendMessageResponse
 from flask_pydantic import validate
 
-from services.guidance_wrapper import GuidanceWrapper
-
-app = Flask(__name__)
-
+from pyris.services.guidance_wrapper import GuidanceWrapper
 
 @app.route("/send-message", methods=["POST"])
 @validate()
@@ -20,7 +17,3 @@ def send_message(body: SendMessageRequest):
         usedModel=body.preferredModel,
         message=SendMessageResponse.Message(content=guidance.query()),
     )
-
-
-if __name__ == "__main__":
-    app.run(debug=True)
